Changed fields to be consistent with database
